requirements_project.txt
pyodbc
psycopg2-binary
pandas
matplotlib
sqlalchemy